+++ shared/r2-shared-swift/Toolkit/Media Type/MediaType.swift
@@ -1,12 +1,7 @@
 //
-//  MediaType.swift
-//  r2-shared-swift
-//
-//  Created by Mickaël Menu on 10/04/2020.
-//
 //  Copyright 2020 Readium Foundation. All rights reserved.
-//  Use of this source code is governed by a BSD-style license which is detailed
-//  in the LICENSE file present in the project repository where this source code is maintained.
+//  Use of this source code is governed by the BSD-style license
+//  available in the top-level LICENSE file of the project.
 //
 
 import Foundation
@@ -22,7 +17,32 @@ import Foundation
 /// media type, for example `application/atom+xml;profile=opds-catalog` for an OPDS 1 catalog.
 ///
 /// Specification: https://tools.ietf.org/html/rfc6838
-public struct MediaType: Equatable, Hashable {
+///
+/// You can register new media types in an extension, for example:
+///
+/// ```
+/// extension MediaType {
+///     var doc: MediaType {
+///         MediaType("application/msword", name: "Word Document", fileExtension: "doc")!
+///     }
+/// }
+/// ```
+public struct MediaType: Hashable, Loggable {
+    
+    /// The string representation of this media type.
+    public var string: String {
+        let params = parameters
+            .map { "\($0.key)=\($0.value)" }
+            .sorted()
+            .joined(separator: ";")
+        return "\(type)/\(subtype)\(params.isEmpty ? "" : ";\(params)")"
+    }
+    
+    /// A human readable name identifying the media type, which may be presented to the user.
+    public let name: String?
+    
+    /// The default file extension to use for this media type.
+    public let fileExtension: String?
     
     /// The type component, e.g. `application` in `application/epub+zip`.
     public let type: String
@@ -32,14 +52,6 @@ public struct MediaType: Equatable, Hashable {
     
     /// The parameters in the media type, such as `charset=utf-8`.
     public let parameters: [String: String]
-    
-    /// The string representation of this media type.
-    public var string: String {
-        let params = parameters.map { "\($0.key)=\($0.value)" }
-            .sorted()
-            .joined(separator: ";")
-        return "\(type)/\(subtype)\(params.isEmpty ? "" : ";\(params)")"
-    }
     
     /// Structured syntax suffix, e.g. `+zip` in `application/epub+zip`
     ///
@@ -57,11 +69,24 @@ public struct MediaType: Equatable, Hashable {
         parameters["charset"].flatMap { String.Encoding(charset: $0) }
     }
 
-    public init?(_ string: String) {
+    /// Returns the canonical version of this media type, if it is known.
+    ///
+    /// This is useful to find the name and file extension of a known media type, or to get the canonical media type
+    /// from an alias. For example, `application/x-cbz` is an alias of the canonical `application/vnd.comicbook+zip`.
+    ///
+    /// Non-significant parameters are also discarded.
+    public var canonicalized: MediaType {
+        MediaType.of(mediaType: string) ?? self
+    }
+
+    public init?(_ string: String, name: String? = nil, fileExtension: String? = nil) {
         guard !string.isEmpty else {
             return nil
         }
-        
+    
+        self.name = name
+        self.fileExtension = fileExtension
+    
         // Grammar: https://tools.ietf.org/html/rfc2045#section-5.1
         let components = string.split(separator: ";")
             .map { $0.trimmingCharacters(in: .whitespaces) }
@@ -142,7 +167,7 @@ public struct MediaType: Equatable, Hashable {
     /// Returns whether this media type and `other` are the same, ignoring parameters that are not
     /// in both media types.
     public func matches(_ other: String?) -> Bool {
-        return matches(other.flatMap(MediaType.init))
+        return matches(other.flatMap { MediaType($0) })
     }
     
     /// Returns whether this media type matches any of the `others` media types.
@@ -218,70 +243,72 @@ public struct MediaType: Equatable, Hashable {
         matchesAny(.lcpProtectedAudiobook, .lcpProtectedPDF)
     }
     
-    /// Returns whether this media type is declared in the Document Types section of the app's main bundle.
+    /// Returns whether this media type is declared in the Document Types section of the app's main
+    /// bundle.
     public var isSupportedDocumentType: Bool {
-        DocumentTypes.main.supportsMediaType(string)
+        return DocumentTypes.main.supportsMediaType(string)
+            || DocumentTypes.main.supportsFileExtension(fileExtension)
     }
 
     
     // MARK: Known Media Types
     
-    public static let aac = MediaType("audio/aac")!
-    public static let acsm = MediaType("application/vnd.adobe.adept+xml")!
-    public static let aiff = MediaType("audio/aiff")!
-    public static let avi = MediaType("video/x-msvideo")!
+    public static let aac = MediaType("audio/aac", fileExtension: "aac")!
+    public static let acsm = MediaType("application/vnd.adobe.adept+xml", name: "Adobe Content Server Message", fileExtension: "acsm")!
+    public static let aiff = MediaType("audio/aiff", fileExtension: "aiff")!
+    public static let avi = MediaType("video/x-msvideo", fileExtension: "avi")!
     public static let binary = MediaType("application/octet-stream")!
-    public static let bmp = MediaType("image/bmp")!
-    public static let cbz = MediaType("application/vnd.comicbook+zip")!
-    public static let css = MediaType("text/css")!
-    public static let divina = MediaType("application/divina+zip")!
-    public static let divinaManifest = MediaType("application/divina+json")!
-    public static let epub = MediaType("application/epub+zip")!
-    public static let gif = MediaType("image/gif")!
-    public static let gz = MediaType("application/gzip")!
-    public static let javascript = MediaType("text/javascript")!
-    public static let jpeg = MediaType("image/jpeg")!
-    public static let html = MediaType("text/html")!
-    public static let json = MediaType("application/json")!
-    public static let lcpProtectedAudiobook = MediaType("application/audiobook+lcp")!
-    public static let lcpProtectedPDF = MediaType("application/pdf+lcp")!
-    public static let lcpLicenseDocument = MediaType("application/vnd.readium.lcp.license.v1.0+json")!
+    public static let bmp = MediaType("image/bmp", fileExtension: "bmp")!
+    public static let cbz = MediaType("application/vnd.comicbook+zip", name: "Comic Book Archive", fileExtension: "cbz")!
+    public static let css = MediaType("text/css", fileExtension: "css")!
+    public static let divina = MediaType("application/divina+zip", name: "Digital Visual Narratives", fileExtension: "divina")!
+    public static let divinaManifest = MediaType("application/divina+json", name: "Digital Visual Narratives", fileExtension: "json")!
+    public static let epub = MediaType("application/epub+zip", name: "EPUB", fileExtension: "epub")!
+    public static let gif = MediaType("image/gif", fileExtension: "gif")!
+    public static let gz = MediaType("application/gzip", fileExtension: "gz")!
+    public static let javascript = MediaType("text/javascript", fileExtension: "js")!
+    public static let jpeg = MediaType("image/jpeg", fileExtension: "jpg")!
+    public static let html = MediaType("text/html", fileExtension: "html")!
+    public static let json = MediaType("application/json", fileExtension: "json")!
+    public static let lcpProtectedAudiobook = MediaType("application/audiobook+lcp", name: "LCP Protected Audiobook", fileExtension: "lcpa")!
+    public static let lcpProtectedPDF = MediaType("application/pdf+lcp", name: "LCP Protected PDF", fileExtension: "lcpdf")!
+    public static let lcpLicenseDocument = MediaType("application/vnd.readium.lcp.license.v1.0+json", name: "LCP License", fileExtension: "lcpl")!
     public static let lcpStatusDocument = MediaType("application/vnd.readium.license.status.v1.0+json")!
-    public static let lpf = MediaType("application/lpf+zip")!
-    public static let mp3 = MediaType("audio/mpeg")!
-    public static let mpeg = MediaType("video/mpeg")!
-    public static let ncx = MediaType("application/x-dtbncx+xml")!
-    public static let ogg = MediaType("audio/ogg")!
-    public static let ogv = MediaType("video/ogg")!
+    public static let lpf = MediaType("application/lpf+zip", fileExtension: "lpf")!
+    public static let mp3 = MediaType("audio/mpeg", fileExtension: "mp3")!
+    public static let mpeg = MediaType("video/mpeg", fileExtension: "mpeg")!
+    public static let ncx = MediaType("application/x-dtbncx+xml", fileExtension: "ncx")!
+    public static let ogg = MediaType("audio/ogg", fileExtension: "oga")!
+    public static let ogv = MediaType("video/ogg", fileExtension: "ogv")!
     public static let opds1 = MediaType("application/atom+xml;profile=opds-catalog")!
     public static let opds1Entry = MediaType("application/atom+xml;type=entry;profile=opds-catalog")!
     public static let opds2 = MediaType("application/opds+json")!
     public static let opds2Publication = MediaType("application/opds-publication+json")!
     public static let opdsAuthentication = MediaType("application/opds-authentication+json")!
-    public static let opus = MediaType("audio/opus")!
-    public static let otf = MediaType("font/otf")!
-    public static let pdf = MediaType("application/pdf")!
-    public static let png = MediaType("image/png")!
-    public static let readiumAudiobook = MediaType("application/audiobook+zip")!
-    public static let readiumAudiobookManifest = MediaType("application/audiobook+json")!
-    public static let readiumWebPub = MediaType("application/webpub+zip")!
-    public static let readiumWebPubManifest = MediaType("application/webpub+json")!
-    public static let smil = MediaType("application/smil+xml")!
-    public static let svg = MediaType("image/svg+xml")!
-    public static let text = MediaType("text/plain")!
-    public static let tiff = MediaType("image/tiff")!
-    public static let ttf = MediaType("font/ttf")!
-    public static let w3cWPUBManifest = MediaType("application/x.readium.w3c.wpub+json")!  // non-existent
-    public static let wav = MediaType("audio/wav")!
-    public static let webmAudio = MediaType("audio/webm")!
-    public static let webmVideo = MediaType("video/webm")!
-    public static let webp = MediaType("image/webp")!
-    public static let woff = MediaType("font/woff")!
-    public static let woff2 = MediaType("font/woff2")!
-    public static let xhtml = MediaType("application/xhtml+xml")!
-    public static let xml = MediaType("application/xml")!
-    public static let zab = MediaType("application/x.readium.zab+zip")!  // non-existent
-    public static let zip = MediaType("application/zip")!
+    public static let opus = MediaType("audio/opus", fileExtension: "opus")!
+    public static let otf = MediaType("font/otf", fileExtension: "otf")!
+    public static let pdf = MediaType("application/pdf", name: "PDF", fileExtension: "pdf")!
+    public static let png = MediaType("image/png", fileExtension: "png")!
+    public static let readiumAudiobook = MediaType("application/audiobook+zip", name: "Readium Audiobook", fileExtension: "audiobook")!
+    public static let readiumAudiobookManifest = MediaType("application/audiobook+json", name: "Readium Audiobook", fileExtension: "json")!
+    public static let readiumWebPub = MediaType("application/webpub+zip", name: "Readium Web Publication", fileExtension: "webpub")!
+    public static let readiumWebPubManifest = MediaType("application/webpub+json", name: "Readium Web Publication", fileExtension: "json")!
+    public static let smil = MediaType("application/smil+xml", fileExtension: "smil")!
+    public static let svg = MediaType("image/svg+xml", fileExtension: "svg")!
+    public static let text = MediaType("text/plain", fileExtension: "txt")!
+    public static let tiff = MediaType("image/tiff", fileExtension: "tiff")!
+    public static let ttf = MediaType("font/ttf", fileExtension: "ttf")!
+    public static let w3cWPUBManifest = MediaType("application/x.readium.w3c.wpub+json", name: "Web Publication", fileExtension: "json")!  // non-existent
+    public static let wav = MediaType("audio/wav", fileExtension: "wav")!
+    public static let webmAudio = MediaType("audio/webm", fileExtension: "webm")!
+    public static let webmVideo = MediaType("video/webm", fileExtension: "webm")!
+    public static let webp = MediaType("image/webp", fileExtension: "webp")!
+    public static let woff = MediaType("font/woff", fileExtension: "woff")!
+    public static let woff2 = MediaType("font/woff2", fileExtension: "woff2")!
+    public static let xhtml = MediaType("application/xhtml+xml", fileExtension: "xhtml")!
+    public static let xml = MediaType("application/xml", fileExtension: "xml")!
+    public static let zab = MediaType("application/x.readium.zab+zip", name: "Zipped Audio Book", fileExtension: "zab")!  // non-existent
+    public static let zip = MediaType("application/zip", fileExtension: "zip")!
     
     static let readiumPositions = MediaType("application/vnd.readium.position-list+json")!
     static let readiumContentProtection = MediaType("application/vnd.readium.content-protection+json")!
@@ -301,15 +328,6 @@ public struct MediaType: Equatable, Hashable {
     public static func ~= (pattern: MediaType, value: MediaType) -> Bool {
         return pattern.matches(value)
     }
-
-    @available(*, unavailable, renamed: "readiumAudiobook")
-    public static var audiobook: MediaType { readiumAudiobook }
-    @available(*, unavailable, renamed: "readiumAudiobookManifest")
-    public static var audiobookManifest: MediaType { readiumAudiobookManifest }
-    @available(*, unavailable, renamed: "readiumWebPub")
-    public static var webpub: MediaType { readiumWebPub }
-    @available(*, unavailable, renamed: "readiumWebPubManifest")
-    public static var webpubManifest: MediaType { readiumWebPubManifest }
 
 }